DEV-36: PR changes move my components to another folder and add new component
