+++ src/components/BlogPosts/MaxWidthContent.scss
@@ -1,0 +1,4 @@
+.max-width-content {
+  width: 1200px;
+  margin: 0px auto;
+}
