+++ src/components/Search.scss
