+++ src/pages/Blog/Blog.tsx
@@ -5,7 +5,8 @@ import Pagination from '../../components/Pagination';
 import './Blog.scss';
 import PostsCardList from '../../components/BlogPosts/PostCardList';
 import { mockPosts } from '../../components/BlogPosts/data';
-import SearchWidget from '../../components/BlogSearch/Search';
+import SearchWidget from '../../components/Search';
+import MaxWidthContent from '../../components/BlogPosts/MaxWidthContent';
 
 const Blog = () => {
   const [featurePost, ...posts] = mockPosts;
@@ -13,20 +14,22 @@ const Blog = () => {
   return (
     <>
       <BlogHeader title="Post Title" subtitle="Post Subtitle" />
-      <Row justify="center" gutter={24}>
-        <Col span={14}>
-          <PostsCardList posts={posts} featurePost={featurePost} />
-          <Pagination
-            totalItems={51}
-            pageSize={10}
-            currentPage={page}
-            onPageChange={onPageChange}
-          />
-        </Col>
-        <Col span={7}>
-          <SearchWidget />
-        </Col>
-      </Row>
+      <MaxWidthContent>
+        <Row justify="center" gutter={24}>
+          <Col span={16}>
+            <PostsCardList posts={posts} featurePost={featurePost} />
+            <Pagination
+              totalItems={51}
+              pageSize={10}
+              currentPage={page}
+              onPageChange={onPageChange}
+            />
+          </Col>
+          <Col span={8}>
+            <SearchWidget />
+          </Col>
+        </Row>
+      </MaxWidthContent>
     </>
   );
 };
